test(e2e): the dashboard specs seed a backdated cycle through one helper
Three dashboard specs each carried their own copy of the same seeding flow:
register an owner, walk onboarding's step-1 picker to an explicit
last-period-start date rather than the today-3 that completeOnboardingIfPresent
hardcodes, then backdate the older cycle starts through the API, which is the
only path with no past-date bound. The copies had already begun to drift — two
returned nothing, one returned the anchor date the assertions need — and each
restated the explicit Origin header a direct API call must carry, so a fourth
spec would have had to rediscover that requirement instead of inheriting it.

registerAndOnboardWithStartDaysAgo and markCycleStartViaAPI now live in
stats-helpers, beside the calendar-driven markCycleStart they are the seeding
counterpart to, together with the isoToday the seeded dates are computed from.
Behaviour is unchanged: no locator, no assertion and no product code moved, and
the three specs run green over plain HTTP and over the TLS posture where the
Origin header is what keeps the CSRF middleware from answering 403.
